Changes for treeview selection,boolean,meshing
Actors in 3d window are now highlighted when selected in model treeview. Removed two unnecessary arrays during boolean process. Changed function to set sphere refinment to set the region to be the maxedgesize if outside sphere region and not already set. This gives better mesh at boundary of sphere region and rest of mesh