Superbuild externals set up
Externals are set up for both superbuild on and off. If on, Externals are built and placed in toplevel directory specified SV_EXTERNALS_TOPLEVEL_DIR. If off, externals are automatically picked up from SV_EXTERNALS_TOPLEVEL_DIR. If this directory doesnt exist or the libraries arent there, system ones are used. On install, if SV_ENABLE_DISTRIBUTION is on, external libs are all installed under one External/lib directory. If SV_ENABLE_DISTRIBUTION is off, the whole Externals folder is installed, which can be used to link to by developers.