BUG: Fix to problem with finding tcl
To fix the problem, a statement was added in front of the find_path and find_library functions in FindTCL.cmake. If TCL_DIR is defined and exists, then the INCLUDE_PATHS and LIBRARIES are unset and searched for again. Otherwise, the current values hold. Find files for TclSh and Wish were added as well