BUG: Fixes to new external build system.
ExternConfig.cmake file location is selected directory location for Externals that have them (i.e. vtk). Must select directories for externals if SV_USE_SYSTEM is on for external or is SV_SUPERBUILD is off. Externals will be automatically detected for build if SV_SUPERBUILD is off and SV_EXTERNALS_USE_TOPLEVEL_DIR is on.