DEV: Integrated vtkSV source
This removes vtk filters in Common Geometry and replaces with vtkSV filters
This also adds much of the rest of the vtkSV code base; however Testing
is left out, the core CMake files to build vtkSV independently are left out
and the module.cmake files to build the libs as vtk modules are left out.
In the future, development on vtkSV should take place in its own repo,
ideally located underneat the simvasuclar repo. If changes are made,
they should be pushed to the vtkSV repo and then pushed to the simvascular
repo. No changes should be made directly to the files in the simvascular
under vtkSV